ENG-4106: add missing files
